Feature: moved "doctor" command into new top level reserved command structure of "ahoy config ..." as "ahoy config validate"
Updated lots of tests, divided commands into separate files.
