+++ v2/testdata/.env.test
@@ -1,0 +1,2 @@
+TEST=value
+ANOTHER_VAR=another_value
